backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
click==8.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.22
starlette==0.50.0
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
requests==2.31.0
google-genai>=1.73.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pytest>=8.3.0
httpx>=0.28.0